// src/app/models/features.ts
export interface features {

    id : number;
    desarrollo : String;
    desarrollador: String;
    zona: String;
    precioMax :number;
    precioMin: number;
    tipo: String;
    medidaMin :number;
    medidaMax: number;
    medidas: string;
    apartado: number;
    tipoenganche: number;
    enganche: number;
    enganche2: number;
    enganche3: string;
    financiamiento: String;
    mantenimiento : String;
    entrega: String;
    entrega2 : any;
    descripcion : String;
    meses : number;
    imagenes:  String;
    src : String;

}